Fix current timestamp bug
In the previous version, I was too eager to reuse
the trick of rstriping 0s and .s from the value.
As a result, if a timestamp ended with 0, we would
remove the 0s and would set the wrong timestamp.

As a result, we could get a current timestamp from
the workflow and convert it back with the same
workflow... but get an incorrect date.

Funnily enough, the GIF screen-capture that is in
the README used to show that bug. We fix the GIF
now as well.