Switch to Python 3 intepretor
A kind contributor sent a pull request to do most of the
change to Python 3. What was left was to be explicit in
the workflow itself (info.plist) to refer to the Python 3
interpretor and either configure the PYTHONPATH there, or
to use explicit relative imports.

This version was verified locally to still work. Hopefully
it also fixes the workflow for people who have seen the
workflow break when macOS removed /usr/bin/python pointing
to a Python 2.7 interpretor.

Fixes #9.